backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow>=2.20.0
librosa>=0.10.1
numpy>=1.26.0
matplotlib>=3.8.2
Pillow>=10.1.0
scipy>=1.11.4
soundfile>=0.12.1
locust>=2.17.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.23
# SQLite is built into Python - no additional driver needed!
alembic>=1.12.1
python-dotenv>=1.0.0

